tdengine/requirements.txt
taospy[ws]
pandas
sqlalchemy